+++ get_types/get_long_double_upper.c
@@ -64,7 +64,7 @@ static char *get_fraction_str(long double frac_part, unsigned int precision, cha
 	i = 0;
 	CH_NULL(frac_part_str = ft_strnew(precision + 1));
 	frac_part_str[i++] = '.';
-	while (i < precision + 1)
+	while (i < (int)(precision + 1))
 	{
 		frac_part *= 10;
 		temp = (long)(frac_part);
@@ -132,7 +132,7 @@ static void round(long double *int_part, long double *frac_part, unsigned int pr
 	{
 		i = 0;
 		div = 0.5;
-		while (i < precision)
+		while (i < (int)precision)
 		{
 			div /= 10;
 			i++;
